requirements.txt
streamlit==1.28.1
streamlit-option-menu==0.3.6
numpy==1.24.3
torch==2.0.1
plotly==5.17.0
keras==2.13.1
Pillow==10.0.0
matplotlib==3.7.2
tensorflow==2.13.0